feat: daily activity for schedule view in separated file
